Fixing of minor typos and commas
Do we write Pull-Request or Pull Request? In this document it was almost consistently written with the dash, but github writes it without one.

Also in line 38, I changed 'organizations' to 'organization'. Is this right? I feel like it refers to the organization devonfw here, or is this more like a general description of how different organizations can have multiple repositories?